recetas/requirements.txt
# Django Framework
Django==4.2.8

# Django REST Framework
djangorestframework==3.14.0

# CORS Headers para Django
django-cors-headers==4.3.1

# TensorFlow para el modelo BERT
tensorflow==2.15.0

# Transformers de Hugging Face para BERT
transformers==4.36.2

# NumPy para operaciones numéricas (compatible con TensorFlow 2.15)
numpy==1.26.4

# Scikit-learn para el label encoder
scikit-learn==1.3.2

# Pillow para manejo de imágenes (opcional, por si se añade funcionalidad)
Pillow==10.1.0

# Requests para llamadas HTTP (útil para testing)
requests==2.31.0

# Python-dotenv para variables de entorno
python-dotenv==1.0.0

# Gunicorn para producción (opcional)
gunicorn==21.2.0

# Whitenoise para servir archivos estáticos en producción
whitenoise==6.6.0

# H5py para inspeccionar archivos de modelo
h5py==3.10.0
